+++ tests/test_tensor/test_belief_propagation/test_d2bp.py
@@ -1,11 +1,7 @@
 import pytest
 
 import quimb.tensor as qtn
-from quimb.experimental.belief_propagation.d2bp import (
-    compress_d2bp,
-    contract_d2bp,
-    sample_d2bp,
-)
+import quimb.tensor.belief_propagation as qbp
 
 
 @pytest.mark.parametrize("damping", [0.0, 0.1])
@@ -16,7 +12,7 @@ def test_contract(damping, dtype, diis):
     # normalize exactly
     peps /= (peps.H @ peps) ** 0.5
     info = {}
-    N_ap = contract_d2bp(
+    N_ap = qbp.contract_d2bp(
         peps, damping=damping, diis=diis, info=info, progbar=True
     )
     assert info["converged"]
@@ -29,7 +25,7 @@ def test_tree_exact(dtype, local_convergence):
     psi = qtn.TN_rand_tree(20, 3, 2, dtype=dtype, seed=42)
     norm2 = psi.H @ psi
     info = {}
-    norm2_bp = contract_d2bp(
+    norm2_bp = qbp.contract_d2bp(
         psi, info=info, local_convergence=local_convergence, progbar=True
     )
     assert info["converged"]
@@ -46,7 +42,7 @@ def test_compress(damping, dtype, diis):
     peps_c1 = peps.compress_all(max_bond=2)
     info = {}
     peps_c2 = peps.copy()
-    compress_d2bp(
+    qbp.compress_d2bp(
         peps_c2,
         max_bond=2,
         damping=damping,
@@ -67,7 +63,7 @@ def test_sample(dtype):
     peps = qtn.PEPS.rand(3, 4, 3, seed=42, dtype=dtype)
     # normalize exactly
     peps /= (peps.H @ peps) ** 0.5
-    config, peps_config, omega = sample_d2bp(peps, seed=42, progbar=True)
+    config, peps_config, omega = qbp.sample_d2bp(peps, seed=42, progbar=True)
     assert all(ix in config for ix in peps.site_inds)
     assert 0.0 < omega < 1.0
     assert peps_config.outer_inds() == ()
@@ -75,7 +71,7 @@ def test_sample(dtype):
     ptotal = 0.0
     nrepeat = 4
     for _ in range(nrepeat):
-        _, peps_config, _ = sample_d2bp(peps, seed=42, progbar=True)
+        _, peps_config, _ = qbp.sample_d2bp(peps, seed=42, progbar=True)
         ptotal += abs(peps_config.contract()) ** 2
 
     # check we are doing better than random guessing