+++ tests/test_tensor/test_belief_propagation/test_l1bp.py
@@ -2,8 +2,7 @@ import pytest
 
 import quimb as qu
 import quimb.tensor as qtn
-from quimb.experimental.belief_propagation.l1bp import contract_l1bp
-from quimb.experimental.belief_propagation.d2bp import contract_d2bp
+import quimb.tensor.belief_propagation as qbp
 
 
 @pytest.mark.parametrize("dtype", ["float32", "complex64"])
@@ -13,7 +12,7 @@ def test_contract_tree_exact(dtype, local_convergence, normalize):
     tn = qtn.TN_rand_tree(10, 3, seed=42, dtype=dtype)
     Z_ex = tn.contract()
     info = {}
-    Z_bp = contract_l1bp(
+    Z_bp = qbp.contract_l1bp(
         tn,
         info=info,
         normalize=normalize,
@@ -31,7 +30,7 @@ def test_contract_loopy_approx(dtype, damping, diis):
     tn = qtn.TN2D_rand(3, 4, 5, dtype=dtype, dist="uniform")
     Z_ex = tn.contract()
     info = {}
-    Z_bp = contract_l1bp(
+    Z_bp = qbp.contract_l1bp(
         tn,
         damping=damping,
         diis=diis,
@@ -50,7 +49,7 @@ def test_contract_double_loopy_approx(dtype, damping, update):
     tn = peps.H & peps
     Z_ex = tn.contract()
     info = {}
-    Z_bp1 = contract_l1bp(
+    Z_bp1 = qbp.contract_l1bp(
         tn,
         damping=damping,
         update=update,
@@ -60,7 +59,7 @@ def test_contract_double_loopy_approx(dtype, damping, update):
     assert info["converged"]
     assert Z_bp1 == pytest.approx(Z_ex, rel=0.3)
     # compare with 2-norm BP on the peps directly
-    Z_bp2 = contract_d2bp(peps)
+    Z_bp2 = qbp.contract_d2bp(peps)
     assert Z_bp1 == pytest.approx(Z_bp2, rel=5e-5)
 
 
@@ -94,7 +93,7 @@ def test_contract_tree_triple_sandwich_exact(dtype):
     tn = bra.H | op | ket
     Z_ex = tn.contract()
     info = {}
-    Z_bp = contract_l1bp(tn, info=info, progbar=True)
+    Z_bp = qbp.contract_l1bp(tn, info=info, progbar=True)
     assert info["converged"]
     assert Z_ex == pytest.approx(Z_bp, rel=5e-6)
 
@@ -120,7 +119,7 @@ def test_contract_tree_triple_sandwich_loopy_approx(dtype, damping, diis):
     tn = ket.H | G_ket
     Z_ex = tn.contract()
     info = {}
-    Z_bp = contract_l1bp(
+    Z_bp = qbp.contract_l1bp(
         tn,
         damping=damping,
         diis=diis,
@@ -134,7 +133,7 @@ def test_contract_tree_triple_sandwich_loopy_approx(dtype, damping, diis):
 def test_contract_cluster_approx():
     tn = qtn.TN2D_classical_ising_partition_function(8, 8, 0.4, h=0.2)
     f_ex = qu.log(tn.contract())
-    f_bp = qu.log(contract_l1bp(tn))
+    f_bp = qu.log(qbp.contract_l1bp(tn))
     assert f_bp == pytest.approx(f_ex, rel=0.3)
     cluster_tags = []
     for i in range(0, 8, 2):
@@ -147,7 +146,7 @@ def test_contract_cluster_approx():
             cluster_tags.append(cluster_tag)
     info = {}
     f_bp2 = qu.log(
-        contract_l1bp(tn, site_tags=cluster_tags, info=info, progbar=True)
+        qbp.contract_l1bp(tn, site_tags=cluster_tags, info=info, progbar=True)
     )
     assert info["converged"]
     assert f_bp == pytest.approx(f_ex, rel=0.1)
@@ -161,7 +160,7 @@ def test_mps():
     psiG = psi.copy()
     psiG.gate_(qu.pauli("X"), 5, contract=True)
     expec = psi.H & psiG
-    O = contract_l1bp(
+    O = qbp.contract_l1bp(
         expec,
         site_tags=[f"I{i}" for i in range(L)],
     )